GitStatsRipoff/requirements.txt
astroid==2.15.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dill==0.3.7
emoji==2.7.0
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.27
isort==5.12.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
matplotlib==3.7.2
mccabe==0.7.0
numpy==1.25.2
packaging==23.1
pexpect==4.8.0
Pillow==10.0.0
pip-autoremove==0.10.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
PyInputPlus==0.2.12
pylint==2.17.5
pyparsing==3.0.9
PySimpleValidate==0.2.12
python-dateutil==2.8.2
six==1.16.0
smmap==5.0.0
stdiomask==0.0.6
termcolor==2.3.0
tomlkit==0.12.1
wcwidth==0.2.6
wrapt==1.15.0
yaspin==2.3.0
